backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai>=1.50.0
pinecone>=5.0.0
langchain>=0.3.0
pypdf>=4.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
